+++ clang/docs/ReleaseNotes.rst
@@ -362,6 +362,9 @@ Bug Fixes to C++ Support
 - Fixed a Clang regression in C++20 mode where unresolved dependent call expressions were created inside non-dependent contexts (#GH122892)
 - Clang now emits the ``-Wunused-variable`` warning when some structured bindings are unused
   and the ``[[maybe_unused]]`` attribute is not applied. (#GH125810)
+- Function parameter packs and constant (non-type) template parameter packs
+  declared with an array or function type are now correctly adjusted to be of
+  the corresponding decayed pointer type.
 
 Bug Fixes to AST Handling
 ^^^^^^^^^^^^^^^^^^^^^^^^^